modification de 32 en 64 dans la struct

+++ include/whirlpool.h
@@ -545,7 +545,7 @@ static const uint64_t C7[256] = {
 
 typedef struct s_whirlpool_ctx
 {
-	uint32_t state[8];  // H0 to H7
+	uint64_t state[8];  // H0 to H7
 	uint64_t bitlen;    // Total length in bits
 	uint8_t buffer[64]; // 512 bits (64 bytes) buffer
 	size_t buffer_len;  // Actual length of the buffer
